--- a/tests/php-unit-tests/phpunit.xml.dist
+++ b/tests/php-unit-tests/phpunit.xml.dist
@@ -20,6 +20,7 @@
 >
 
   <php>
+    <ini name="memory_limit" value="512M"/>
     <ini name="error_reporting" value="E_ALL"/>
     <ini name="display_errors" value="On"/>
     <ini name="log_errors" value="On"/>