--- a/tests/php-unit-tests/src/BaseTestCase/ItopDataTestCase.php
+++ b/tests/php-unit-tests/src/BaseTestCase/ItopDataTestCase.php
@@ -92,6 +92,7 @@ abstract class ItopDataTestCase extends ItopTestCase
 	 */
 	public static function tearDownAfterClass(): void
 	{
+		\UserRights::FlushPrivileges();
 		parent::tearDownAfterClass();
 	}