--- a/tests/php-unit-tests/src/BaseTestCase/ItopTestCase.php
+++ b/tests/php-unit-tests/src/BaseTestCase/ItopTestCase.php
@@ -24,6 +24,8 @@ abstract class ItopTestCase extends TestCase
 	const TEST_LOG_DIR = 'test';
 	static $DEBUG_UNIT_TEST = false;
 
+	private $aConfigOriginalValues = [];
+
 	/**
 	 * Override the default value to disable the backup of globals in case of tests run in a separate process
 	 */
@@ -112,6 +114,13 @@ abstract class ItopTestCase extends TestCase
 				CMDBSource::Query('ROLLBACK');
 			}
 			throw new MySQLTransactionNotClosedException('Some DB transactions were opened but not closed ! Fix the code by adding ROLLBACK or COMMIT statements !', []);
+		}
+
+		if (count($this->aConfigOriginalValues) > 0) {
+			$oConfig = \utils::GetConfig();
+			foreach ($this->aConfigOriginalValues as $sKey => $value) {
+				$oConfig->Set($sKey, $value);
+			}
 		}
 	}
 
@@ -211,6 +220,27 @@ abstract class ItopTestCase extends TestCase
 		$sId = str_replace(' ', '_', $sId);
 
 		return $sId;
+	}
+
+	/**
+	 * Facade vor utils::GetConfig()->Set()
+	 *
+	 * @param string $sKey
+	 * @param mixed $value
+	 *
+	 * @return void
+	 */
+	public function ConfigSet(string $sKey, $value)
+	{
+		if (!array_key_exists($sKey, $this->aConfigOriginalValues)) {
+			$this->aConfigOriginalValues[$sKey] = $value;
+		}
+		\utils::GetConfig()->Set($sKey, $value);
+	}
+
+	public function GetConfigClone()
+	{
+		return clone \utils::GetConfig();
 	}
 
 	/**