--- a/tests/php-unit-tests/unitary-tests/application/Session/SessionTest.php
+++ b/tests/php-unit-tests/unitary-tests/application/Session/SessionTest.php
@@ -7,8 +7,6 @@ use Combodo\iTop\Test\UnitTest\ItopTestCase;
 
 /**
  * @runTestsInSeparateProcesses
- * @preserveGlobalState disabled
- * @backupGlobals disabled
  */
 class SessionTest extends ItopTestCase
 {