--- a/tests/php-unit-tests/unitary-tests/application/ThemeHandlerTest.php
+++ b/tests/php-unit-tests/unitary-tests/application/ThemeHandlerTest.php
@@ -7,9 +7,6 @@ use FindStylesheetObject;
 use ThemeHandler;
 
 /**
- * @runTestsInSeparateProcesses
- * @preserveGlobalState disabled
- * @backupGlobals disabled
  * @covers ThemeHandler
  */
 class ThemeHandlerTest extends ItopTestCase