--- a/tests/php-unit-tests/unitary-tests/application/UI/Base/Layout/NavigationMenuTest.php
+++ b/tests/php-unit-tests/unitary-tests/application/UI/Base/Layout/NavigationMenuTest.php
@@ -9,9 +9,6 @@ use Combodo\iTop\Test\UnitTest\ItopDataTestCase;
 
 /**
  *
- * @runTestsInSeparateProcesses
- * @preserveGlobalState disabled
- * @backupGlobals disabled
  * Class NavigationMenuTest
  *
  * @package UI\Base\Layout