# Intention
The PHPUnit test suite is slow to execute.
This is an issue on the CI pipeline.
This is a **stopper** for developpers who cannot afford running the whole test suite neither regularly nor occasionaly!

I have the intuition that PHPUnit tests could be significantly accelerated (10 times faster), by many ways, and wanted to experiment that for good.

# Benchmarks

## Evaluate the overhead of @runTestsInSeparateProcesses

Those results show that, for some tests, most of the time is spent in parsing iTop files and initializing the data structures.

The parsing of PHP files can be improved by the mean of opcache (requires an explicit declaration in CLI), while the initialization of iTop data structures relies on opcache/data and still needs some heavy deserialization of its data.

### A test that does noting but assertTrue(true)
Note : still, the test was based in ItopDataTestCase
![image](https://github.com/Combodo/iTop/assets/6113279/ab9c8b58-7380-4e26-a16c-a2ad8e30a503)

### ExpressionEvaluateTest::testVariousExpressions (83 cases)
Initially, with process isolation: 56"
![image](https://github.com/Combodo/iTop/assets/6113279/6a27fc36-7c16-4502-9be2-a3c60a6b0754)

Without the process isolation: 0.7" (plus: setupBeforeClass seems not taken into account)
![image](https://github.com/Combodo/iTop/assets/6113279/38754095-728d-4fe0-ae0a-16e99e729ba8)


## The whole suite

| Case | Before | After |
|--------|--------|--------|
| CI Jenkins, opcache CLI enabled | 5'10" | 2'15" |
| PC Romain, Windows, PHP 7.4.33, xdebug.mode=develop | 20'30" | 6'52" |
| PC Romain, Windows, PHP 7.4.33, xdebug.mode=off | 17'33" | 4'37" |
| PC Pierre, Windows, PHP 8.1.21, opcache enabled, xdebug_mode=develop |  | 6'7 | 
| PC Pierre, Windows, PHP 8.1.21, opcache enabled, xdebug_mode=off | 19'36 | 3'55 | 

## Slowest test case
(based on Romain's PC)

### xdebug.mode=develop

Out of 99 test classes, 12 have a duration exceeding 10"
Those slow tests represent 82% of the whole time spent.

Tests during more than 10" 
| Test | Duration | % |
|--------|--------|--------|
|CriterionConversionTest|76,1|18%|
|DictionariesConsistency|49,3|12%|
|ItopTicketTest|46,8|11%|
|SearchFormTest|36|9%|
|CliResetSessionTest|22,2|5%|
|BulkChangeTest|21,7|5%|
|UserRightsTest|21,6|5%|
|TransactionsTest|18,8|5%|
|RestTest|13,1|3%|
|DBSearchTest|11,9|3%|
|BulkChangeExtKeyTest|10,9|3%|
|ExceptionLogTest|10,8|3%|
|Sous-total|339,2|82%|
|Total|412|100%|

### xdebug.mode=off

| Test | Duration | % |
|--------|--------|--------|
|DictonariesConsistencyTest|44,4|16%|
|CliResetSessionTest|21,6|8%|
|BulkChangeExtKeyTest|21,2|8%|
|BulkChangeTest|17,9|6%|
|UserRightsTest|16,5|6%|
|RestTest|13,2|5%|
|ItopTicketTest|11,5|4%|
|Sous-total|146,3|53%|
|Total|277|100%|

MySQL queries performed during the tests:

| Operation | Occurrences | Acc. Duration |
|--------|--------|--------|
|DELETE FROM|7176|2"|
|INSERT INTO|9111|2,3"|
|SELECT COUNT|2599|1"|
|SELECT DISCTINCT|16925|10"|
|SET NAMES|88|0,032"|
|UPDATE|355|0,115"|
|COMMIT|4582|0,77"|
|ROLLBACK|627|1,1"|
|START TRANSACTION|5208|0,23"|
|Total|46671|17,5"


# Solutions
## Do not isolate tests execution (whenever possible)
@runTestsInSeparateProcesses have been removed in many cases
Sometimes, the process isolation has been moved at the test level instead of the class level.

This is the most effective way of boosting the execution of the test suite.

Anyhow, it assumes that a tests leaves the place clean: any change in static data should be reset.

Example : Dict::SetLanguage requires a reset to the default in the end. 

## Eliminate redundancy in tests
### RestTest
Most of the tests here were doing 3 to 5 calls to the API, resulting in a redundancy.

## Test through HTTP only if necessary
### RestTest
The list_operations command is the only case in which HTPP is used.
For other APIs; the test is performed directly by instanciating CoreServices, thus bypassing the job done in rest.php.

## Todo
- [X] Fix tests not working on Windows
- [X] Enable opcache.enable_cli on the CI : divided the execution time by two on the CI !
- [X] Perform a benchmark on another system
- [X] Remove any false positive on the CI
- [X] Remove any false positive on Windows
- [ ] See how opcache could be enable on Windows (failing to do so until now)
- [x] Document each forbidden pattern (polluter) and its countermeasure (cleanup on you own, use a facade or let the framework to the job for you)
- [x] Finalize the test that detects polluting tests
- Additional performance improvements :
  - [ ] See if some object creation (login user, organization) could be handled in setupBeforeClass rather than doing so in setup.
  - [X] Do a final pass on tests still isolated... do we have a clue now?
  - [ ] Analyze the slowest test classes: is there still a quick win?
- [X] Do a report of (already performed) benchmarks (DB operations vs PHP vs datamodel startup vs Time to rollback vs ...)
- [X] Rebase on develop (PR on datamodel dependant tests should cause conflicts)
- [x] Rebase on support/3.0
- [x] Merge on support/3.1
- [x] Merge on develop
- Experiments
  - [x] Clarify if setupBeforeClass is taken into account in the duration. If not, then reconsider the whole statistics (could be wrong)
  - [X] Find a way to list test classes through PHPUnit CLI => --list-tests
  - [X] Write a script that will execute each test class separately
  - [ ] Finalize or drop the tool to find which test is not compatible with a given test (still hardcoded: the whole list of tests)
  - [ ] Experiment a mean to detect polluters based on reality (analyse static data changes by the mean of the PHP reflection API)