@piRGoif 2023-08-07
You mean, "Boost PHPUnit execution" instead of boot I think :)